MCUToken/server/requirements.txt
omegaconf
pyserial
matplotlib
scikit-learn
scikit-learn-intelex
tqdm
numpy
pandas
seaborn
